url_shortner_server/requirements.txt
asgiref==3.5.2
# backports.zoneinfo==0.2.1
bootstrap4==0.1.0
coverage==6.5.0
Django==4.1.3
django-cors-headers==3.13.0
flake8==5.0.4
mccabe==0.7.0
mysqlclient==2.1.1
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv
pytz==2021.1
requests
sqlparse==0.4.2
